requirements.txt
streamlit
openai-whisper
torch
ffmpeg-python
indic-transliteration